Add TestLinkSetup class. Starting refactor of test link setup
